ci: Some more refactor
